Move authentication mixin to a namespace that does not conflict with existing node/auth functionality. Dedupe code constants to fix notification tests.
