+++ core/node/notifications/service.go
@@ -9,12 +9,13 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 
 	"github.com/river-build/river/core/config"
-	"github.com/river-build/river/core/node/auth"
+	"github.com/river-build/river/core/node/authentication"
 	. "github.com/river-build/river/core/node/base"
 	"github.com/river-build/river/core/node/crypto"
 	"github.com/river-build/river/core/node/events"
 	"github.com/river-build/river/core/node/infra"
 	"github.com/river-build/river/core/node/logging"
+
 	"github.com/river-build/river/core/node/nodes"
 	notificationssync "github.com/river-build/river/core/node/notifications/sync"
 	"github.com/river-build/river/core/node/notifications/types"
@@ -25,7 +26,7 @@ import (
 
 type (
 	Service struct {
-		auth.AuthServiceMixin
+		authentication.AuthServiceMixin
 		notificationsConfig config.NotificationsConfig
 		onChainConfig       crypto.OnChainConfiguration
 		userPreferences     UserPreferencesStore
@@ -103,7 +104,7 @@ func (s *Service) GetSettings(
 	ctx context.Context,
 	req *connect.Request[GetSettingsRequest],
 ) (*connect.Response[GetSettingsResponse], error) {
-	userID := ctx.Value(UserIDCtxKey{}).(common.Address)
+	userID := ctx.Value(authentication.UserIDCtxKey{}).(common.Address)
 	if userID == (common.Address{}) {
 		return nil, RiverError(Err_INVALID_ARGUMENT, "Invalid user id")
 	}
@@ -147,7 +148,7 @@ func (s *Service) SetSettings(
 	ctx context.Context,
 	req *connect.Request[SetSettingsRequest],
 ) (*connect.Response[SetSettingsResponse], error) {
-	userID := ctx.Value(UserIDCtxKey{}).(common.Address)
+	userID := ctx.Value(authentication.UserIDCtxKey{}).(common.Address)
 	if userID == (common.Address{}) {
 		return nil, RiverError(Err_INVALID_ARGUMENT, "Invalid user id")
 	}
@@ -175,7 +176,7 @@ func (s *Service) SetDmGdmSettings(
 		gdm = msg.GetGdmGlobal()
 	)
 
-	userID := ctx.Value(UserIDCtxKey{}).(common.Address)
+	userID := ctx.Value(authentication.UserIDCtxKey{}).(common.Address)
 	if userID == (common.Address{}) {
 		return nil, RiverError(Err_INVALID_ARGUMENT, "Invalid user id")
 	}
@@ -198,7 +199,7 @@ func (s *Service) SetSpaceSettings(
 		value        = msg.GetValue()
 	)
 
-	userID := ctx.Value(UserIDCtxKey{}).(common.Address)
+	userID := ctx.Value(authentication.UserIDCtxKey{}).(common.Address)
 	if userID == (common.Address{}) {
 		return nil, RiverError(Err_INVALID_ARGUMENT, "Invalid user id")
 	}
@@ -225,7 +226,7 @@ func (s *Service) SetDmChannelSetting(
 		value = msg.GetValue()
 	)
 
-	userID := ctx.Value(UserIDCtxKey{}).(common.Address)
+	userID := ctx.Value(authentication.UserIDCtxKey{}).(common.Address)
 	if userID == (common.Address{}) {
 		return nil, RiverError(Err_INVALID_ARGUMENT, "Invalid user id")
 	}
@@ -256,7 +257,7 @@ func (s *Service) SetGdmChannelSetting(
 		value = msg.GetValue()
 	)
 
-	userID := ctx.Value(UserIDCtxKey{}).(common.Address)
+	userID := ctx.Value(authentication.UserIDCtxKey{}).(common.Address)
 	if userID == (common.Address{}) {
 		return nil, RiverError(Err_INVALID_ARGUMENT, "Invalid user id")
 	}
@@ -287,7 +288,7 @@ func (s *Service) SetSpaceChannelSettings(
 		value = msg.GetValue()
 	)
 
-	userID := ctx.Value(UserIDCtxKey{}).(common.Address)
+	userID := ctx.Value(authentication.UserIDCtxKey{}).(common.Address)
 	if userID == (common.Address{}) {
 		return nil, RiverError(Err_INVALID_ARGUMENT, "Invalid user id")
 	}
@@ -321,7 +322,9 @@ func (s *Service) SubscribeWebPush(
 		}
 	)
 
-	userID := ctx.Value(UserIDCtxKey{}).(common.Address)
+	logging.FromCtx(ctx).Infow("context", "ctx", ctx, "userId", ctx.Value(authentication.UserIDCtxKey{}))
+
+	userID := ctx.Value(authentication.UserIDCtxKey{}).(common.Address)
 	if userID == (common.Address{}) {
 		return nil, RiverError(Err_INVALID_ARGUMENT, "Invalid user id")
 	}
@@ -354,7 +357,7 @@ func (s *Service) UnsubscribeWebPush(
 		}
 	)
 
-	userID := ctx.Value(UserIDCtxKey{}).(common.Address)
+	userID := ctx.Value(authentication.UserIDCtxKey{}).(common.Address)
 	if userID == (common.Address{}) {
 		return nil, RiverError(Err_INVALID_ARGUMENT, "Invalid user id")
 	}
@@ -376,7 +379,7 @@ func (s *Service) SubscribeAPN(
 ) (*connect.Response[SubscribeAPNResponse], error) {
 	var (
 		msg         = req.Msg
-		userID      = ctx.Value(UserIDCtxKey{}).(common.Address)
+		userID      = ctx.Value(authentication.UserIDCtxKey{}).(common.Address)
 		deviceToken = msg.GetDeviceToken()
 		environment = msg.GetEnvironment()
 		pushVersion = msg.GetPushVersion()
@@ -407,7 +410,7 @@ func (s *Service) UnsubscribeAPN(
 	var (
 		msg         = req.Msg
 		deviceToken = msg.GetDeviceToken()
-		userID      = ctx.Value(UserIDCtxKey{}).(common.Address)
+		userID      = ctx.Value(authentication.UserIDCtxKey{}).(common.Address)
 	)
 	if len(deviceToken) == 0 {
 		return nil, RiverError(Err_INVALID_ARGUMENT, "Invalid APN device token")
